+++ pyhap/camera.py
@@ -213,7 +213,7 @@ NO_SRTP = b'\x01\x01\x02\x02\x00\x03\x00'
 
 FFMPEG_CMD = (
     # pylint: disable=bad-continuation
-    'ffmpeg -re -f avfoundation -i 0:0 -threads 0 '
+    'ffmpeg -re -f avfoundation -framerate {fps} -i 0:0 -threads 0 '
     '-vcodec libx264 -an -pix_fmt yuv420p -r {fps} -f rawvideo -tune zerolatency '
     '-vf scale={width}:{height} -b:v {v_max_bitrate}k -bufsize {v_max_bitrate}k '
     '-payload_type 99 -ssrc {v_ssrc} -f rtp '
@@ -222,6 +222,8 @@ FFMPEG_CMD = (
     'localrtcpport={v_port}&pkt_size=1378'
 )
 '''Template for the ffmpeg command.'''
+
+logger = logging.getLogger(__name__)
 
 
 class Camera(Accessory):
@@ -310,7 +312,7 @@ class Camera(Accessory):
                 codec = AUDIO_CODEC_TYPES['AACELD']
                 bitrate = AUDIO_CODEC_PARAM_BIT_RATE_TYPES['VARIABLE']
             else:
-                logging.warning('Unsupported codec %s', param_type)
+                logger.warning('Unsupported codec %s', param_type)
                 continue
 
             param_samplerate = codec_param['samplerate']
@@ -321,7 +323,7 @@ class Camera(Accessory):
             elif param_samplerate == 24:
                 samplerate = AUDIO_CODEC_PARAM_SAMPLE_RATE_TYPES['KHZ_24']
             else:
-                logging.warning('Unsupported sample rate %s', param_samplerate)
+                logger.warning('Unsupported sample rate %s', param_samplerate)
                 continue
 
             param_tlv = tlv.encode(AUDIO_CODEC_PARAM_TYPES['CHANNEL'], b'\x01',
@@ -332,7 +334,7 @@ class Camera(Accessory):
             configs += tlv.encode(SUPPORTED_AUDIO_CODECS_TAG, config_tlv)
 
         if not has_supported_codec:
-            logging.warning('Client does not support any audio codec that iOS supports.')
+            logger.warning('Client does not support any audio codec that iOS supports.')
 
             codec = AUDIO_CODEC_TYPES['OPUS']
             bitrate = AUDIO_CODEC_PARAM_BIT_RATE_TYPES['VARIABLE']
@@ -408,7 +410,6 @@ class Camera(Accessory):
 
         :type options: ``dict``
         """
-        self.streaming_status = STREAMING_STATUS['AVAILABLE']
         self.has_srtp = options.get('srtp', False)
         self.start_stream_cmd = options.get('start_stream_cmd', FFMPEG_CMD)
 
@@ -423,22 +424,52 @@ class Camera(Accessory):
         super().__init__(*args, **kwargs)
 
         self.add_preload_service('Microphone')
-        management = self.add_preload_service('CameraRTPStreamManagement')
-        management.configure_char('StreamingStatus',
-                                  getter_callback=self._get_streaming_status)
-        management.configure_char('SupportedRTPConfiguration',
-                                  value=self.get_supported_rtp_config(
-                                                options.get('srtp', False)))
-        management.configure_char('SupportedVideoStreamConfiguration',
-                                  value=self.get_supported_video_stream_config(
-                                                options['video']))
-        management.configure_char('SupportedAudioStreamConfiguration',
-                                  value=self.get_supported_audio_stream_config(
-                                                options['audio']))
-        management.configure_char('SelectedRTPStreamConfiguration',
-                                  setter_callback=self.set_selected_stream_configuration)
-        management.configure_char('SetupEndpoints',
-                                  setter_callback=self.set_endpoints)
+        self._streaming_status = []
+        self._management = []
+        self._setup_stream_management(options)
+
+    @property
+    def streaming_status(self):
+        """For backwards compatibility."""
+        return self._streaming_status[0]
+
+    def _setup_stream_management(self, options):
+        """Create stream management."""
+        stream_count = options.get("stream_count", 1)
+        for stream_idx in range(stream_count):
+            self._management.append(self._create_stream_management(stream_idx, options))
+            self._streaming_status.append(STREAMING_STATUS["AVAILABLE"])
+
+    def _create_stream_management(self, stream_idx, options):
+        """Create a stream management service."""
+        management = self.add_preload_service("CameraRTPStreamManagement")
+        management.configure_char(
+            "StreamingStatus",
+            getter_callback=lambda: self._get_streaming_status(stream_idx),
+        )
+        management.configure_char(
+            "SupportedRTPConfiguration",
+            value=self.get_supported_rtp_config(options.get("srtp", False)),
+        )
+        management.configure_char(
+            "SupportedVideoStreamConfiguration",
+            value=self.get_supported_video_stream_config(options["video"]),
+        )
+        management.configure_char(
+            "SupportedAudioStreamConfiguration",
+            value=self.get_supported_audio_stream_config(options["audio"]),
+        )
+        management.configure_char(
+            "SelectedRTPStreamConfiguration",
+            setter_callback=self.set_selected_stream_configuration,
+        )
+        management.configure_char(
+            "SetupEndpoints",
+            setter_callback=lambda value: self.set_endpoints(
+                value, stream_idx=stream_idx
+            ),
+        )
+        return management
 
     async def _start_stream(self, objs, reconfigure):  # pylint: disable=unused-argument
         """Start or reconfigure video streaming for the given session.
@@ -483,20 +514,25 @@ class Camera(Accessory):
             video_rtp_param = video_objs.get(VIDEO_TYPES['RTP_PARAM'])
             if video_rtp_param:
                 video_rtp_param_objs = tlv.decode(video_rtp_param)
-                # TODO: Optionals, handle the case where they are missing
-                opts['v_ssrc'] = struct.unpack('<I',
-                    video_rtp_param_objs.get(
-                        RTP_PARAM_TYPES['SYNCHRONIZATION_SOURCE']))[0]
-                opts['v_payload_type'] = \
-                    video_rtp_param_objs.get(RTP_PARAM_TYPES['PAYLOAD_TYPE'])
-                opts['v_max_bitrate'] = struct.unpack('<H',
-                    video_rtp_param_objs.get(RTP_PARAM_TYPES['MAX_BIT_RATE']))[0]
-                opts['v_rtcp_interval'] = struct.unpack('<f',
-                    video_rtp_param_objs.get(RTP_PARAM_TYPES['RTCP_SEND_INTERVAL']))[0]
-                opts['v_max_mtu'] = video_rtp_param_objs.get(RTP_PARAM_TYPES['MAX_MTU'])
+                if RTP_PARAM_TYPES['SYNCHRONIZATION_SOURCE'] in video_rtp_param_objs:
+                    opts['v_ssrc'] = struct.unpack('<I',
+                        video_rtp_param_objs.get(
+                            RTP_PARAM_TYPES['SYNCHRONIZATION_SOURCE']))[0]
+                if RTP_PARAM_TYPES['PAYLOAD_TYPE'] in video_rtp_param_objs:
+                    opts['v_payload_type'] = \
+                        video_rtp_param_objs.get(RTP_PARAM_TYPES['PAYLOAD_TYPE'])
+                if RTP_PARAM_TYPES['MAX_BIT_RATE'] in video_rtp_param_objs:
+                    opts['v_max_bitrate'] = struct.unpack('<H',
+                        video_rtp_param_objs.get(RTP_PARAM_TYPES['MAX_BIT_RATE']))[0]
+                if RTP_PARAM_TYPES['RTCP_SEND_INTERVAL'] in video_rtp_param_objs:
+                    opts['v_rtcp_interval'] = struct.unpack('<f',
+                        video_rtp_param_objs.get(RTP_PARAM_TYPES['RTCP_SEND_INTERVAL']))[0]
+                if RTP_PARAM_TYPES['MAX_MTU'] in video_rtp_param_objs:
+                    opts['v_max_mtu'] = video_rtp_param_objs.get(RTP_PARAM_TYPES['MAX_MTU'])
 
         if audio_tlv:
             audio_objs = tlv.decode(audio_tlv)
+
             opts['a_codec'] = audio_objs[AUDIO_TYPES['CODEC']]
             audio_codec_param_objs = tlv.decode(
                                         audio_objs[AUDIO_TYPES['CODEC_PARAM']])
@@ -526,25 +562,28 @@ class Camera(Accessory):
         session_objs = tlv.decode(objs[SELECTED_STREAM_CONFIGURATION_TYPES['SESSION']])
         session_id = UUID(bytes=session_objs[SETUP_TYPES['SESSION_ID']])
         session_info = self.sessions[session_id]
+        stream_idx = session_info['stream_idx']
 
         opts.update(session_info)
         success = await self.reconfigure_stream(session_info, opts) if reconfigure \
             else await self.start_stream(session_info, opts)
 
         if success:
-            self.streaming_status = STREAMING_STATUS['STREAMING']
+            self._streaming_status[stream_idx] = STREAMING_STATUS['STREAMING']
         else:
-            logging.error('[%s] Faled to start/reconfigure stream, deleting session.',
-                          session_id)
+            logger.error(
+                '[%s] Failed to start/reconfigure stream, deleting session.',
+                session_id
+            )
             del self.sessions[session_id]
-            self.streaming_status = STREAMING_STATUS['AVAILABLE']
+            self._streaming_status[stream_idx] = STREAMING_STATUS['AVAILABLE']
 
-    def _get_streaming_status(self):
+    def _get_streaming_status(self, stream_idx):
         """Get the streaming status in TLV format.
 
         Called when iOS reads the StreaminStatus ``Characteristic``.
         """
-        return tlv.encode(b'\x01', self.streaming_status, to_base64=True)
+        return tlv.encode(b'\x01', self._streaming_status[stream_idx], to_base64=True)
 
     async def _stop_stream(self, objs):
         """Stop the stream for the specified session.
@@ -558,16 +597,20 @@ class Camera(Accessory):
         session_id = UUID(bytes=session_objs[SETUP_TYPES['SESSION_ID']])
 
         session_info = self.sessions.get(session_id)
+        stream_idx = session_info['stream_idx']
 
         if not session_info:
-            logging.error('Requested to stop stream for session %s, but no '
-                          'such session was found', session_id)
+            logger.error(
+                'Requested to stop stream for session %s, but no '
+                'such session was found',
+                session_id
+            )
             return
 
         await self.stop_stream(session_info)
         del self.sessions[session_id]
 
-        self.streaming_status = STREAMING_STATUS['AVAILABLE']
+        self._streaming_status[stream_idx] = STREAMING_STATUS['AVAILABLE']
 
     def set_selected_stream_configuration(self, value):
         """Set the selected stream configuration.
@@ -580,17 +623,17 @@ class Camera(Accessory):
         :param value: base64-encoded selected configuration in TLV format
         :type value: ``str``
         """
-        logging.debug('set_selected_stream_config - value - %s', value)
+        logger.debug('set_selected_stream_config - value - %s', value)
 
         objs = tlv.decode(value, from_base64=True)
         if SELECTED_STREAM_CONFIGURATION_TYPES['SESSION'] not in objs:
-            logging.error('Bad request to set selected stream configuration.')
+            logger.error('Bad request to set selected stream configuration.')
             return
 
         session = tlv.decode(objs[SELECTED_STREAM_CONFIGURATION_TYPES['SESSION']])
 
         request_type = session[b'\x02'][0]
-        logging.debug('Set stream config request: %d', request_type)
+        logger.debug('Set stream config request: %d', request_type)
         if request_type == 1:
             job = functools.partial(self._start_stream, reconfigure=False)
         elif request_type == 0:
@@ -598,12 +641,17 @@ class Camera(Accessory):
         elif request_type == 4:
             job = functools.partial(self._start_stream, reconfigure=True)
         else:
-            logging.error('Unknown request type %d', request_type)
+            logger.error('Unknown request type %d', request_type)
             return
 
         self.driver.add_job(job, objs)
 
-    def set_endpoints(self, value):
+    def set_streaming_available(self, stream_idx):
+        """Send an update to the controller that streaming is available."""
+        self._streaming_status[stream_idx] = STREAMING_STATUS["AVAILABLE"]
+        self._management[stream_idx].get_characteristic("StreamingStatus").notify()
+
+    def set_endpoints(self, value, stream_idx=None):
         """Configure streaming endpoints.
 
         Called when iOS sets the SetupEndpoints ``Characteristic``. The endpoint
@@ -612,6 +660,9 @@ class Camera(Accessory):
         :param value: The base64-encoded stream session details in TLV format.
         :param value: ``str``
         """
+        if stream_idx is None:
+            stream_idx = 0
+
         objs = tlv.decode(value, from_base64=True)
         session_id = UUID(bytes=objs[SETUP_TYPES['SESSION_ID']])
 
@@ -640,16 +691,18 @@ class Camera(Accessory):
         audio_master_key = audio_info_objs[SETUP_SRTP_PARAM['MASTER_KEY']]
         audio_master_salt = audio_info_objs[SETUP_SRTP_PARAM['MASTER_SALT']]
 
-        logging.debug('Received endpoint configuration:'
-                      '\nsession_id: %s\naddress: %s\nis_ipv6: %s'
-                      '\ntarget_video_port: %s\ntarget_audio_port: %s'
-                      '\nvideo_crypto_suite: %s\nvideo_srtp: %s'
-                      '\naudio_crypto_suite: %s\naudio_srtp: %s',
-                      session_id, address, is_ipv6, target_video_port, target_audio_port,
-                      video_crypto_suite,
-                      to_base64_str(video_master_key + video_master_salt),
-                      audio_crypto_suite,
-                      to_base64_str(audio_master_key + audio_master_salt))
+        logger.debug(
+            'Received endpoint configuration:'
+            '\nsession_id: %s\naddress: %s\nis_ipv6: %s'
+            '\ntarget_video_port: %s\ntarget_audio_port: %s'
+            '\nvideo_crypto_suite: %s\nvideo_srtp: %s'
+            '\naudio_crypto_suite: %s\naudio_srtp: %s',
+            session_id, address, is_ipv6, target_video_port, target_audio_port,
+            video_crypto_suite,
+            to_base64_str(video_master_key + video_master_salt),
+            audio_crypto_suite,
+            to_base64_str(audio_master_key + audio_master_salt)
+        )
 
         # Configure the SetupEndpoints response
 
@@ -688,6 +741,7 @@ class Camera(Accessory):
 
         self.sessions[session_id] = {
             'id': session_id,
+            'stream_idx': stream_idx,
             'address': address,
             'v_port': target_video_port,
             'v_srtp_key': to_base64_str(video_master_key + video_master_salt),
@@ -697,9 +751,7 @@ class Camera(Accessory):
             'a_ssrc': audio_ssrc
         }
 
-        self.get_service('CameraRTPStreamManagement')\
-            .get_characteristic('SetupEndpoints')\
-            .set_value(response_tlv)
+        self._management[stream_idx].get_characteristic('SetupEndpoints').set_value(response_tlv)
 
     async def stop(self):
         """Stop all streaming sessions."""
@@ -765,24 +817,30 @@ class Camera(Accessory):
         :return: True if and only if starting the stream command was successful.
         :rtype: ``bool``
         """
-        logging.debug('[%s] Starting stream with the following parameters: %s',
-                      session_info['id'], stream_config)
+        logger.debug(
+            '[%s] Starting stream with the following parameters: %s',
+            session_info['id'],
+            stream_config
+        )
 
         cmd = self.start_stream_cmd.format(**stream_config).split()
-        logging.debug('Executing start stream command: "%s"', ' '.join(cmd))
+        logger.debug('Executing start stream command: "%s"', ' '.join(cmd))
         try:
             process = await asyncio.create_subprocess_exec(*cmd,
                     stdout=asyncio.subprocess.DEVNULL,
                     stderr=asyncio.subprocess.PIPE,
                     limit=1024)
         except Exception as e:  # pylint: disable=broad-except
-            logging.error('Failed to start streaming process because of error: %s', e)
+            logger.error('Failed to start streaming process because of error: %s', e)
             return False
 
         session_info['process'] = process
 
-        logging.info('[%s] Started stream process - PID %d',
-                     session_info['id'], process.pid)
+        logger.info(
+            '[%s] Started stream process - PID %d',
+            session_info['id'],
+            process.pid
+        )
 
         return True
 
@@ -800,20 +858,22 @@ class Camera(Accessory):
         session_id = session_info['id']
         ffmpeg_process = session_info.get('process')
         if ffmpeg_process:
-            logging.info('[%s] Stopping stream.', session_id)
+            logger.info('[%s] Stopping stream.', session_id)
             try:
                 ffmpeg_process.terminate()
                 _, stderr = await asyncio.wait_for(
                     ffmpeg_process.communicate(), timeout=2.0)
-                logging.debug('Stream command stderr: %s', stderr)
+                logger.debug('Stream command stderr: %s', stderr)
             except asyncio.TimeoutError:
-                logging.error('Timeout while waiting for the stream process '
-                              'to terminate. Trying with kill.')
+                logger.error(
+                    'Timeout while waiting for the stream process '
+                    'to terminate. Trying with kill.'
+                )
                 ffmpeg_process.kill()
                 await ffmpeg_process.wait()
-            logging.debug('Stream process stopped.')
+            logger.debug('Stream process stopped.')
         else:
-            logging.warning('No process for session ID %s', session_id)
+            logger.warning('No process for session ID %s', session_id)
 
     async def reconfigure_stream(self, session_info, stream_config):
         """Reconfigure the stream so that it uses the given ``stream_config``.