+++ tests/test_camera.py
@@ -6,6 +6,7 @@ from pyhap import camera
 
 
 _OPTIONS = {
+    "stream_count": 4,
     "video": {
         "codec": {
             "profiles": [
@@ -91,7 +92,7 @@ def test_set_selected_stream_start_stop(mock_driver):
     process_mock = Mock()
 
     # Mock for asyncio.create_subprocess_exec
-    async def subprocess_exec(*args, **kwargs):
+    async def subprocess_exec(*args, **kwargs):  # pylint: disable=unused-argument
         process_mock.id = 42
         process_mock.communicate = communicate
         process_mock.wait = wait
@@ -105,6 +106,7 @@ def test_set_selected_stream_start_stop(mock_driver):
 
     session_info = {
         'id': session_id,
+        'stream_idx': 0,
         'address': '192.168.1.114',
         'v_port': 50483,
         'v_srtp_key': '2JZgpMkwWUH8ahUtzp8VThtBmbk26hCPJqeWpYDR',